requirements.txt
attrs==24.2.0
blinker==1.9.0
click==8.1.7
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
load-dotenv==0.1.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mistune==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.10.0
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
